requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
braintree==4.19.0
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-crontab==0.7.1
django-import-export==3.2.0
django-localflavor==4.0
django-modeltranslation==0.18.9
django-rosetta==0.9.9
et-xmlfile==1.1.0
flower==1.2.0
fonttools==4.39.4
html5lib==1.1
humanize==4.6.0
idna==3.4
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
polib==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
python-decouple==3.8
python-slugify==8.0.1
python-stdnum==1.18
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.1
typing_extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.2
